consolidate adapter related tests to inference_with_adapters_test.go
